refactor(sign up form component): organize Sign up page components
organized and refactored the components on Sign Up Page to improve the overall structure and
readability of the code
