+++ src/pages/signup/index.js
@@ -1,7 +1,7 @@
 import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 
-import SignUpForm from "@/components/SignUpForm";
+import SignUpForm from "@/components/signUp/SignUpForm";
 import SittingImg from "@/components/SittingImg";
 
 const SingUp = () => {
